requirements.txt
openai
python-dotenv
requests
pip
clang
gremlinpython
aiohttp
async_timeout